@@ -5,15 +5,11 @@ import java.io.InputStreamReader
 import java.net.HttpURLConnection
 import java.net.URL
 
-
 class TranslateUtil {
     companion object {
-        private const val endpoint =
-            "https://script.google.com/macros/s/AKfycbzEAPpoqrAgkIHyG1-ao2RmW4HxhLJgeVukZ6yBYMp58QvxaWA6eHvzfeBHiUd401SZVQ/exec"
-
         fun translateEnToJa(text: String): String {
             try {
-                val url = URL("$endpoint?text=$text&target=ja")
+                val url = URL("${BuildConfig.translateEndPoint}?text=$text&target=ja")
                 val urlConnection = url.openConnection() as HttpURLConnection
                 urlConnection.requestMethod = "GET"
                 urlConnection.connect()
@@ -32,7 +28,7 @@ class TranslateUtil {
 
         fun translateJaToEn(text: String): String {
             try {
-                val url = URL("$endpoint?text=$text&target=en")
+                val url = URL("${BuildConfig.translateEndPoint}?text=$text&target=en")
                 val urlConnection = url.openConnection() as HttpURLConnection
                 urlConnection.requestMethod = "GET"
                 urlConnection.connect()
